% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaSeqNormalizer.R
\name{data.frame.rnaSeqNormalizer}
\alias{data.frame.rnaSeqNormalizer}
\title{create object of class rnaSeqNormalizer, initializing with a data.frame}
\usage{
data.frame.rnaSeqNormalizer(tbl)
}
\arguments{
\item{mtx}{a matix in "raw" form}
}
\value{
A normalized matrix with dimensions, row and column names preserved
}
\description{
create object of class rnaSeqNormalizer, initializing with a data.frame
}
