% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaSeqNormalization.R
\docType{methods}
\name{getNormalizedMatrix,rnaSeqNormalization-method}
\alias{getNormalizedMatrix,rnaSeqNormalization-method}
\alias{getNormalizedMatrix}
\title{get the normalized matrix}
\usage{
\S4method{getNormalizedMatrix}{rnaSeqNormalization}(obj)
}
\arguments{
\item{obj}{An object of class rnaSeqNormalization}
}
\description{
get the normalized matrix
}
