% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaSeqNormalizer.R
\docType{methods}
\name{normalize,rnaSeqNormalizer-method}
\alias{normalize,rnaSeqNormalizer-method}
\alias{normalize}
\title{do the normalizer}
\usage{
\S4method{normalize}{rnaSeqNormalizer}(obj)
}
\arguments{
\item{obj}{An object of class rnaSeqNormalizer}
}
\description{
do the normalizer
}
