% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaSeqNormalization.R
\docType{class}
\name{rnaSeqNormalization-class}
\alias{rnaSeqNormalization-class}
\alias{.rnaSeqNormalization}
\alias{rnaSeqNormalization}
\title{rnaSeqNormalization}
\usage{
rnaSeqNormalization(mtx)
}
\arguments{
\item{mtx}{a matix in "raw" form}
}
\value{
A normalized matrix with dimensions, row and column names preserved
}
\description{
Cory, Michael and Max brewed up this method, here packaged up for easy reuse.
}
