% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaSeqNormalizer.R
\docType{class}
\name{rnaSeqNormalizer-class}
\alias{rnaSeqNormalizer-class}
\alias{.rnaSeqNormalizer}
\alias{rnaSeqNormalizer}
\title{rnaSeqNormalizer}
\usage{
rnaSeqNormalizer(mtx)
}
\arguments{
\item{mtx}{a matix in "raw" form}
}
\value{
A normalized matrix with dimensions, row and column names preserved
}
\description{
Cory, Michael and Max brewed up this method, here packaged up for easy reuse.
}
