% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaSeqNormalization.R
\docType{methods}
\name{show,rnaSeqNormalization-method}
\alias{show,rnaSeqNormalization-method}
\alias{show}
\title{get a summary of the objects}
\usage{
\S4method{show}{rnaSeqNormalization}(object)
}
\arguments{
\item{obj}{An object of class rnaSeqNormalization}
}
\description{
get a summary of the objects
}
