% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaSeqNormalizer.R
\docType{methods}
\name{standardizeToGeneSymbolMatrix,rnaSeqNormalizer-method}
\alias{standardizeToGeneSymbolMatrix,rnaSeqNormalizer-method}
\alias{toGeneSymbolMatrix}
\title{convert data.frame to a geneSymbol matrix}
\usage{
\S4method{standardizeToGeneSymbolMatrix}{rnaSeqNormalizer}(obj)
}
\arguments{
\item{obj}{An object of class rnaSeqNormalizer}

\item{method}{A character string specifying statistic to use to resolve the case in which
multiple ENSEMBL identifiers map to the same gene symbol: one of "median", "mean", "sd"}
}
\description{
convert data.frame to a geneSymbol matrix
}
