% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaSeqNormalizer.R
\docType{methods}
\name{getNormalizedMatrix,rnaSeqNormalizer-method}
\alias{getNormalizedMatrix,rnaSeqNormalizer-method}
\alias{getNormalizedMatrix}
\title{get the normalized matrix}
\usage{
\S4method{getNormalizedMatrix}{rnaSeqNormalizer}(obj)
}
\arguments{
\item{obj}{An object of class rnaSeqNormalizer}
}
\description{
get the normalized matrix
}
