% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaSeqNormalizer.R
\name{show,rnaSeqNormalizer-method}
\alias{show,rnaSeqNormalizer-method}
\alias{show}
\title{get a summary of the objects}
\usage{
\S4method{show}{rnaSeqNormalizer}(object)
}
\arguments{
\item{obj}{An object of class rnaSeqNormalizer}
}
\description{
get a summary of the objects
}
